Training fixes: canvas_mse checkpoint selection, post-init seeding, 50 epochs
Investigated exist_decay topology bleed where slot-3 strokes die between
epoch 29-49, leaving gaps in R bowl, 8 top loop, etc. Run 19 dodged this
by being killed at epoch 59. Fixes:

- Checkpoint selection by canvas_mse (not total loss which rewards pruning)
- Post-init seeding: torch/numpy/random seeds set AFTER model creation so
  weight init matches run 19's distribution. Pre-init seeding produced 15%
  worse convergence from a bad initialization.
- 50 epochs (down from 100): canvas_mse plateaus by epoch 46, further
  training only degrades topology via exist_decay bleed.
- uint8 glyph cache: 75GB→19GB, prevents disk-full crashes on Vast.ai.
- Crash detection in deploy monitor: auto-destroys on OSError/OOM/CUDA.
- compare_runs.py: HTML comparison page, auto-opens after sync.
- render_checkpoints_oldviz.py: renders multiple checkpoints in old viz
  style for apples-to-apples comparison.

+++ font_scraper/docker/stroke_model/dataset.py
@@ -116,7 +116,7 @@ class GlyphDataset(Dataset):
             if os.path.exists(cache_path):
                 try:
                     arr = np.load(cache_path)
-                    img_tensor = torch.from_numpy(arr).float().unsqueeze(0)
+                    img_tensor = torch.from_numpy(arr.astype(np.float32)).unsqueeze(0)
                     return img_tensor, char_idx
                 except (EOFError, ValueError):
                     pass  # Corrupt cache file, fall through to render
@@ -137,10 +137,10 @@ class GlyphDataset(Dataset):
         if self.augment:
             img_tensor = self._augment(img_tensor)
 
-        # Cache to disk
+        # Cache to disk as uint8 (binary mask: 0 or 1) — 4x smaller than float32
         if self.cache_dir:
             try:
-                np.save(cache_path, img_arr)
+                np.save(cache_path, img_arr.astype(np.uint8))
             except OSError:
                 pass
